chore: move sign transaction to transaction.ts
